+++ src/scribe/plugins/meeting/__init__.py
@@ -1,0 +1,7 @@
+"""
+Meeting Note Taker Plugin
+"""
+
+from .plugin import MeetingNoteTakerPlugin
+
+__all__ = ['MeetingNoteTakerPlugin']